refactor: refactor chat code
