+++ lib-chat/src/main/resources/cn/wzpmc/dao/ChatDao.xml
@@ -16,6 +16,6 @@
         select `base64` from user where `username`=#{username};
     </select>
     <select id="getCount" resultType="java.lang.Integer">
-        select COUNT(*) from message where recall=0;
+        select COUNT(*) from message;
     </select>
 </mapper>
