+++ web/src/components/mainpage.vue
@@ -17,10 +17,10 @@
 <!--            //使用to来实现路由跳转
             //active-class 当给链接被激活时展示的样式-->
             <router-link :to="'/main/chat?a='+this.username" active-class="active-mainPages"><t-menu-item value="chat"><icon slot="icon" name="chat"/>聊天</t-menu-item></router-link>
-            <router-link :to="'/main/voicechat?a='+this.username" active-class="active-mainPages"><t-menu-item value="voice-chat"><icon slot="icon" name="service"/>语音聊天</t-menu-item></router-link>
+            <!--router-link :to="'/main/voicechat?a='+this.username" active-class="active-mainPages"><t-menu-item value="voice-chat"><icon slot="icon" name="service"/>语音聊天</t-menu-item></router-link>
             <router-link :to="'/main/file?a='+this.username" active-class="active-mainPages"><t-menu-item value="file"><icon slot="icon" name="cloud-upload"/>文件传输</t-menu-item></router-link>
             <router-link to="#" active-class="active-mainPages"><t-menu-item value="game"><icon slot="icon" name="app"/>小♂游戏</t-menu-item></router-link>
-            <router-link to="#" active-class="active-mainPages" ><t-menu-item value="settings"><icon slot="icon" name="setting"/>个人设置</t-menu-item></router-link>
+            <router-link to="#" active-class="active-mainPages" ><t-menu-item value="settings"><icon slot="icon" name="setting"/>个人设置</t-menu-item></router-link-->
           </t-menu>
         </t-aside>
         <t-layout>
